Rename package and response encoder. Include entity encoder in body enconder.
