+++ src/main/scala/com/github/aldtid/developers/connected/application.scala
@@ -1,33 +1,33 @@
 package com.github.aldtid.developers.connected
 
 import com.github.aldtid.developers.connected.controller.DevelopersController
+import com.github.aldtid.developers.connected.encoder.BodyEncoder
+import com.github.aldtid.developers.connected.encoder.implicits._
 import com.github.aldtid.developers.connected.logging.ProgramLog
 import com.github.aldtid.developers.connected.logging.implicits.all._
 import com.github.aldtid.developers.connected.logging.messages._
 import com.github.aldtid.developers.connected.model.Developers
-import com.github.aldtid.developers.connected.model.encoder.ResponseEncoder
-import com.github.aldtid.developers.connected.model.encoder.implicits._
 import com.github.aldtid.developers.connected.model.responses.{Errors, MissingResource}
 
 import cats.Monad
 import cats.effect.Sync
 import cats.implicits._
-import org.http4s.{EntityEncoder, HttpApp, Request, Response}
+import org.http4s.{HttpApp, Request, Response}
 import org.http4s.dsl.Http4sDsl
 import org.typelevel.log4cats.Logger
 
 
 object application {
 
-  def app[F[_] : Sync : Logger : Http4sDsl, L, O : ResponseEncoder](controller: DevelopersController[F])
-                                                                   (implicit pl: ProgramLog[L],
-                                                                    encoder: EntityEncoder[F, O]): HttpApp[F] = {
+  def app[F[_] : Sync : Logger : Http4sDsl, L, O : BodyEncoder](controller: DevelopersController[F])
+                                                               (implicit pl: ProgramLog[L]): HttpApp[F] = {
 
     import pl._
 
     HttpApp[F](request =>
 
       for {
+
         start    <- Sync[F].realTime
         _        <- Logger[F].info(incomingRequest |+| request)
 
@@ -36,6 +36,7 @@ object application {
         end      <- Sync[F].realTime
         latency   = (end - start).toMillis.asLatency
         _        <- Logger[F].info(outgoingResponse |+| response |+| latency)
+
       } yield response
 
     )
@@ -45,10 +46,9 @@ object application {
   def process[F[_] : Monad, L : ProgramLog, O](request: Request[F],
                                                controller: DevelopersController[F])
                                               (implicit dsl: Http4sDsl[F],
-                                               re: ResponseEncoder[O],
-                                               encoder: EntityEncoder[F, O]): F[Response[F]] = {
+                                               be: BodyEncoder[O]): F[Response[F]] = {
 
-    import dsl._, re._
+    import dsl._, be._
 
     request match {
