+++ src/main/scala/com/github/aldtid/developers/connected/encoder/BodyEncoder.scala
@@ -1,9 +1,12 @@
-package com.github.aldtid.developers.connected.model.encoder
+package com.github.aldtid.developers.connected.encoder
 
 import com.github.aldtid.developers.connected.model.responses.{Connection, Errors}
+import org.http4s.EntityEncoder
 
 
-trait ResponseEncoder[O] {
+trait BodyEncoder[O] {
+
+  implicit def entityEncoder[F[_]]: EntityEncoder[F, O]
 
   implicit val connectionEncoder: Encoder[Connection, O]
   implicit val errorsEncoder: Encoder[Errors, O]
