+++ src/main/scala/com/github/aldtid/developers/connected/encoder/Encoder.scala
@@ -1,0 +1,8 @@
+package com.github.aldtid.developers.connected.encoder
+
+
+trait Encoder[A, O] {
+
+  def encode(a: A): O
+
+}
