--- src/main/scala/com/github/aldtid/developers/connected/model/encoder/Encoder.scala
@@ -1,8 +1,0 @@
-package com.github.aldtid.developers.connected.model.encoder
-
-
-trait Encoder[A, O] {
-
-  def encode(a: A): O
-
-}
